[#02][be][user]refactor: 이메일 코드 인증 후 저장 시 ttl 변경 외 스토리지로 네이밍 변경
