+++ back/src/main/java/com/rouby/user/application/service/verification/VerificationEmailCodeStorage.java
@@ -1,11 +1,12 @@
-package com.rouby.user.domain.repository;
+package com.rouby.user.application.service.verification;
 
-import com.rouby.user.application.entity.VerificationEmailCode;
 import java.util.Optional;
 
-public interface VerificationEmailCodeRepository {
+public interface VerificationEmailCodeStorage {
 
   void save(String email, VerificationEmailCode code);
+
+  void verified(String email, VerificationEmailCode code);
 
   Optional<VerificationEmailCode> findByEmail(String email);
