+++ back/src/main/java/com/rouby/user/infrastructure/persistence/redis/RedisVerificationEmailCodeStorage.java
@@ -1,7 +1,7 @@
 package com.rouby.user.infrastructure.persistence.redis;
 
-import com.rouby.user.application.entity.VerificationEmailCode;
-import com.rouby.user.domain.repository.VerificationEmailCodeRepository;
+import com.rouby.user.application.service.verification.VerificationEmailCode;
+import com.rouby.user.application.service.verification.VerificationEmailCodeStorage;
 import java.time.Duration;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
@@ -10,15 +10,20 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 @RequiredArgsConstructor
-public class RedisVerificationEmailCodeRepository implements VerificationEmailCodeRepository {
+public class RedisVerificationEmailCodeStorage implements VerificationEmailCodeStorage {
 
   private static final String KEY = "verification:email:";
   private static final Duration ttl = Duration.ofMinutes(5);
+  private static final Duration verifiedTtl = Duration.ofMinutes(60);
 
   private final RedisTemplate<String, Object> redisTemplate;
 
   public void save(String email, VerificationEmailCode code) {
     redisTemplate.opsForValue().set(buildKey(email), code, ttl);
+  }
+
+  public void verified(String email, VerificationEmailCode code) {
+    redisTemplate.opsForValue().set(buildKey(email), code, verifiedTtl);
   }
 
   public Optional<VerificationEmailCode> findByEmail(String email) {
